feat: 参加者の増減処理をドメインサービス化
